pyproject.toml
[project]
name = "holmes-hydro"
version = "3.3.4"
description = "HOLMES (HydrOLogical Modeling Educationnal Software) is a software developped to teach operational hydrology. It is developed at the university Laval, Québec, Canada."
readme = "README.md"
authors = [
  {name = "Antoine Lefebvre-Brossard", email = "antoinelb@proton.me"}
]
license = {text = "MIT"}
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Education",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Operating System :: OS Independent",
  "Topic :: Scientific/Engineering :: Hydrology",
]
requires-python = ">=3.11"
dependencies = [
    "holmes-rs>=0.2.1",
    "numpy>=2.4.0",
    "polars>=1.36.0",
    "starlette>=0.50.0",
    "uvicorn>=0.36.0",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "httpx>=0.28.0",
    "hypothesis>=6.0.0",
    "isort>=6.0.1",
    "maturin>=1.10.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-playwright>=0.6.2",
    "ruff>=0.13.1",
    "ty>=0.0.1a21",
]

[project.urls]
Homepage = "https://github.com/antoinelb/holmes"
Repository = "https://github.com/antoinelb/holmes"

[tool.uv.sources]
holmes-rs = { path = "./src/holmes-rs", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/holmes"]

[tool.hatch.build.targets.sdist]
include = ["src/", "tests/", "docs/", "pyproject.toml", "README.md"]

[project.scripts]
holmes = "holmes.app:run_server"

[tool.black]
line_length = 79
target-version = ['py313']

[tool.isort]
profile = "black"
src_paths = ["holmes", "holmes_rs", "tests"]

[tool.pytest.ini_options]
minversion = "8.0"
# Order matters: unit tests must run before e2e to avoid event loop pollution from Playwright
testpaths = ["tests/unit", "tests/integration", "tests/e2e"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["src/holmes"]
omit = [
    "*/__pycache__/*",
    "*/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]
precision = 2
show_missing = true

